requirements.txt
absl-py==1.4.0
accelerate==0.21.0
addict==2.4.0
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
basicsr==1.4.2
beautifulsoup4==4.12.2
blendmodes==2022
boltons==23.0.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.7
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
cmake==3.27.5
colorama==0.4.6
colour-science==0.4.4
contourpy==1.1.1
cycler==0.11.0
deprecation==2.1.0
einops==0.4.1
exceptiongroup==1.1.3
facexlib==0.3.0
fastapi==0.94.0
ffmpy==0.3.1
filelock==3.12.4
filterpy==1.4.5
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
ftfy==6.1.1
future==0.18.3
fuzzywuzzy==0.18.0
gdown==4.7.1
gfpgan==1.3.8
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.23.0
google-auth-oauthlib==1.0.0
gradio==3.41.2
gradio_client==0.5.0
grpcio==1.58.0
h11==0.12.0
hepunits==2.3.2
httpcore==0.15.0
httpx==0.24.1
huggingface-hub==0.17.1
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
importlib-resources==6.0.1
inflection==0.5.1
iso8601==2.1.0
Jinja2==3.1.2
jsonmerge==1.8.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kornia==0.6.7
lark==1.1.2
lazy_loader==0.3
Levenshtein==0.25.1
lightning-utilities==0.9.0
lit==16.0.6
llvmlite==0.40.1
lmdb==1.4.1
loguru==0.7.2
lpips==0.1.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
mpld3==0.5.10
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.2.3
open-clip-torch==2.20.0
opencv-python==4.8.0.76
orjson==3.9.7
packaging==23.1
paho-mqtt==2.0.0
pandas==2.1.0
particle==0.23.1
piexif==1.1.3
Pillow==9.5.0
platformdirs==3.10.0
playsound==1.3.0
pocketsphinx==5.0.3
protobuf==3.20.0
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycairo==1.25.1
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
pygame==2.5.2
PyGObject==3.46.0
PyGObject-stubs==2.10.0
pymandel==1.0.12
pyparsing==3.1.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.25.1
python-multipart==0.0.6
pytorch-lightning==1.9.4
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.9.1
realesrgan==0.3.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rpds-py==0.10.3
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scipy==1.11.2
semantic-version==2.10.0
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.5
SpeechRecognition==3.10.4
starlette==0.26.1
sympy==1.12
tb-nightly==2.15.0a20230915
tensorboard-data-server==0.7.1
thefuzz==0.22.1
tifffile==2023.8.30
timm==0.9.2
tokenizers==0.13.3
tomesd==0.1.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1+cu118
torchdiffeq==0.2.3
torchmetrics==1.1.2
torchsde==0.2.5
torchvision==0.15.2+cu118
tqdm==4.66.1
trampoline==0.1.2
transformers==4.30.2
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2
